+++ linen-starter/src/main/AndroidManifest.xml
@@ -20,12 +20,12 @@
         </activity>
 
         <activity
-            android:name=".SettingChannelsActivity"
+            android:name=".settings.SettingChannelsActivity"
             >
         </activity>
 
         <activity
-            android:name=".SettingChannelSourcesActivity"
+            android:name=".settings.SettingChannelSourcesActivity"
             >
         </activity>
