+++ linen-starter/src/main/java/x7c1/linen/settings/SettingChannelsActivity.java
@@ -1,8 +1,10 @@
-package x7c1.linen;
+package x7c1.linen.settings;
 
 import android.os.Bundle;
 import android.view.KeyEvent;
 
+import x7c1.linen.BaseActivity;
+import x7c1.linen.R;
 import x7c1.linen.glue.res.layout.SettingChannelsLayout;
 import x7c1.linen.modern.init.settings.SettingChannelsDelegatee;
 import x7c1.linen.res.layout.SettingChannelsLayoutProvider;
